chore: stabilize microservices stack and fix deployment blockers
- Fix api-gateway routing and service discovery
- Fix station-service hibernate scale errors and schema mismatches
- Fix auth-service startup crash by removing broken wallet repository
- Fix frontend port conflict (moved to 3001)
- Add missing Dockerfiles and Spring Boot entry points for core services
- Stabilize energy-reporting-service compilation issues

+++ backend/billing-service/src/main/java/com/evroaminghub/billing/kafka/SessionEventConsumer.java
@@ -46,7 +46,7 @@ public class SessionEventConsumer {
                     .sessionId(UUID.fromString(sessionId))
                     .userId(UUID.fromString(userId))
                     .energyKwh(energyKwh)
-                    .durationMinutes(durationMinutes)
+                    .durationMinutes((double) durationMinutes)
                     .sessionStartedAt(start)
                     .sessionStoppedAt(stop)
                     .tariffId(defaultTariffId)
